third_party/blink/web_tests/external/wpt/reporting/reporting-api-orders-reports-buffered.https-expected.txt

This is a testharness.js-based test.
[FAIL] Buffered ReportingObserver preserves report generation order
  assert_array_equals: expected property 0 to be "csp-violation" but got "test" (expected array ["csp-violation", "test", "csp-violation", "test", "csp-violation", "test", "csp-violation", "test", "csp-violation", "test", "csp-violation", "test", "csp-violation", "test", "csp-violation", "test", "csp-violation", "test", "csp-violation", "test", …] got ["test", "test", "test", "test", "test", "test", "test", "test", "test", "test", "test", "test", "test", "test", "test", "test", "test", "test", "test", "test", …])
Harness: the test ran to completion.
